requirements.txt
pywhatkit~=5.4
requests~=2.28.1
pandas~=1.5.3
numpy~=1.24.1
mysql-connector-python~=8.0.32
python-dotenv~=0.21.1
python-dotenv

setuptools~=67.2.0
DateTime~=5.0